+++ cdl/core/gui/processor/signal.py
@@ -83,6 +83,16 @@ class SignalProcessor(BaseProcessor):
         self.compute_11(cps.compute_abs, title=_("Absolute value"))
 
     @qt_try_except()
+    def compute_re(self) -> None:
+        """Compute real part"""
+        self.compute_11(cps.compute_re, title=_("Real part"))
+
+    @qt_try_except()
+    def compute_im(self) -> None:
+        """Compute imaginary part"""
+        self.compute_11(cps.compute_im, title=_("Imaginary part"))
+
+    @qt_try_except()
     def compute_log10(self) -> None:
         """Compute Log10"""
         self.compute_11(cps.compute_log10, title="Log10")